% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnsplots-package.R
\docType{package}
\name{cnsplots-package}
\alias{cnsplots}
\alias{cnsplots-package}
\title{cnsplots: scientific plotting tools for R}
\description{
The native R package follows the visual baseline of Python cnsplots 0.5.0
while returning ordinary ggplot2 objects. Package loading has no theme,
option, font, or graphics-device side effects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/faridrashidi/cnsplots}
  \item Report bugs at \url{https://github.com/faridrashidi/cnsplots/issues}
}

}
\author{
\strong{Maintainer}: jarxunlai \email{1927756261@qq.com} (Native R implementation)

Authors:
\itemize{
  \item jarxunlai \email{1927756261@qq.com} (Native R implementation)
  \item Farid Rashidi \email{farid.rsh@gmail.com} (Original cnsplots design and Python implementation) [copyright holder]
}

}
\keyword{internal}
