% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{colour_constants}
\alias{colour_constants}
\alias{RED}
\alias{BLUE}
\alias{GREEN}
\alias{PURPLE}
\alias{ORANGE}
\alias{YELLOW}
\alias{BROWN}
\alias{PINK}
\alias{GRAY}
\alias{VIOLET}
\alias{CHOCOLATE}
\title{Author colour constants}
\format{
A length-one character vector containing a hexadecimal R colour.
}
\usage{
RED

BLUE

GREEN

PURPLE

ORANGE

YELLOW

BROWN

PINK

GRAY

VIOLET

CHOCOLATE
}
\description{
Length-one hexadecimal colour tokens matching the author's Python API.
They can be used anywhere an R colour is accepted.
}
