% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_names}
\alias{palette_names}
\title{List available palettes}
\usage{
palette_names(kind = c("all", "qualitative", "continuous"), details = FALSE)
}
\arguments{
\item{kind}{One of "all", "qualitative", or "continuous".}

\item{details}{Return the palette metadata instead of only names.}
}
\value{
A character vector or metadata data frame.
}
\description{
List available palettes
}
