% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-basic.R
\name{placeholderplot}
\alias{placeholderplot}
\title{Draw a placeholder panel}
\usage{
placeholderplot(description)
}
\arguments{
\item{description}{Text shown in the placeholder caption.}
}
\value{
A ggplot object.
}
\description{
Draw a placeholder panel
}
