% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-distribution.R
\name{qqplot}
\alias{qqplot}
\title{Draw a normal quantile-quantile plot}
\usage{
qqplot(data, x, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Numeric column name.}

\item{...}{Additional named arguments passed to \code{ggplot2::geom_point()}.}
}
\value{
A ggplot object.
}
\description{
Plotting positions use \code{i / (n + 1)}, the default used by the Python
statsmodels implementation. No reference line is added by default.
}
