% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{reset_settings}
\alias{reset_settings}
\title{Restore all cnsplots settings}
\usage{
reset_settings()
}
\value{
Invisibly, the settings in effect before the reset.
}
\description{
Restore all cnsplots settings
}
