% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_colour_map}
\alias{scale_colour_map}
\alias{scale_color_map}
\alias{scale_fill_map}
\title{Use a continuous colour map}
\usage{
scale_colour_map(palette = NULL, ..., na.value = "grey50")

scale_color_map(palette = NULL, ..., na.value = "grey50")

scale_fill_map(palette = NULL, ..., na.value = "grey50")
}
\arguments{
\item{palette}{A continuous palette name.}

\item{...}{Arguments passed to the ggplot2 scale constructor.}

\item{na.value}{Colour used for missing values.}
}
\value{
A ggplot2 scale.
}
\description{
Use a continuous colour map
}
