% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_colour_palette}
\alias{scale_colour_palette}
\alias{scale_color_palette}
\alias{scale_fill_palette}
\title{Use a qualitative palette for colour}
\usage{
scale_colour_palette(palette = NULL, ..., na.value = "grey50")

scale_color_palette(palette = NULL, ..., na.value = "grey50")

scale_fill_palette(palette = NULL, ..., na.value = "grey50")
}
\arguments{
\item{palette}{A qualitative palette name or custom colour vector.}

\item{...}{Arguments passed to the ggplot2 scale constructor.}

\item{na.value}{Colour used for missing values.}
}
\value{
A ggplot2 scale.
}
\description{
Use a qualitative palette for colour
}
