% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-basic.R
\name{scatterplot}
\alias{scatterplot}
\title{Draw a scatter plot}
\usage{
scatterplot(
  data,
  x,
  y,
  s = 7,
  hue = NULL,
  hue_order = NULL,
  palette = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{x, y}{Column names.}

\item{s}{Matplotlib-compatible marker area in points squared.}

\item{hue}{Optional grouping column.}

\item{hue_order}{Optional order of hue levels.}

\item{palette}{Palette name or custom colours.}

\item{...}{Additional arguments passed to \code{ggplot2::geom_point()}.}
}
\value{
A ggplot object.
}
\description{
Draw a scatter plot
}
