% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-components.R
\name{theme_axes}
\alias{theme_axes}
\title{Modify cnsplots axes}
\usage{
theme_axes(x = TRUE, y = TRUE, ticks = TRUE, tick_labels = TRUE, titles = TRUE)
}
\arguments{
\item{x, y}{Whether to display each axis.}

\item{ticks}{Whether to display tick marks.}

\item{tick_labels}{Whether to display tick labels.}

\item{titles}{Whether to display axis titles.}
}
\value{
An incomplete ggplot2 theme patch.
}
\description{
Modify cnsplots axes
}
