% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-components.R
\name{theme_grid}
\alias{theme_grid}
\title{Modify cnsplots grid lines}
\usage{
theme_grid(
  major = c("none", "x", "y", "both"),
  minor = c("none", "x", "y", "both"),
  colour = "grey85",
  linewidth = 0.3
)
}
\arguments{
\item{major, minor}{One of "none", "x", "y", or "both".}

\item{colour}{Grid colour.}

\item{linewidth}{Grid width in points.}
}
\value{
An incomplete ggplot2 theme patch.
}
\description{
Modify cnsplots grid lines
}
