% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-components.R
\name{theme_legend}
\alias{theme_legend}
\title{Modify a cnsplots legend}
\usage{
theme_legend(position = "right", direction = NULL, title = TRUE)
}
\arguments{
\item{position}{A ggplot2 legend position.}

\item{direction}{Optional "horizontal" or "vertical".}

\item{title}{Whether to display the legend title.}
}
\value{
An incomplete ggplot2 theme patch.
}
\description{
Modify a cnsplots legend
}
