% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-components.R
\name{theme_spacing}
\alias{theme_spacing}
\title{Modify cnsplots spacing}
\usage{
theme_spacing(
  plot_margin = c(0, 0, 0, 0),
  panel_spacing = 4,
  legend_spacing = 2
)
}
\arguments{
\item{plot_margin}{Numeric top, right, bottom, and left margins in points.}

\item{panel_spacing}{Panel spacing in points.}

\item{legend_spacing}{Legend spacing in points.}
}
\value{
An incomplete ggplot2 theme patch.
}
\description{
Modify cnsplots spacing
}
